Refactor CLI startup and update agent instructions
