+++ HelloWorldApp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelloWorldApp", "HelloWorldApp\HelloWorldApp.csproj", "{88C31C7C-CC48-4C24-A964-429D2A0EEE73}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinearCongruentGenerator.CLI", "LinearCongruentGenerator.CLI\LinearCongruentGenerator.CLI.csproj", "{88C31C7C-CC48-4C24-A964-429D2A0EEE73}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LinearCongruentGenerator", "LinearCongruentGenerator\LinearCongruentGenerator.csproj", "{B757D739-0577-42C8-9115-EABBDDB6C273}"
 EndProject
