+++ LinearCongruentGenerator.CLI/Program.cs
@@ -1,0 +1,1 @@
+return LinearCongruentGenerator.CLI.CliApp.Run(args);
